Consolidate Pages hexToRgb onto one snippet; prune unreferenced dev PNGs
Item 3 — the client-side hexToRgb emitted into the generated Pages HTML was
copy-pasted into three generators (scripts/site/generate.ts, differences.ts,
xychart-test.ts) and had drifted: differences.ts dropped the `length !== 6`
guard, so a malformed 5-char hex parsed to a wrong colour instead of null.
Extract one canonical snippet in scripts/site/client-color.ts and interpolate
it into all three client <script> templates. (It stays a JS string because it
runs in the browser; the typed server-side primitives remain in
src/shared/color-math.ts.) Generators run clean; property-html-generator and
comparison-differences-sync tests pass.

Item 5 — delete unreferenced committed dev artifacts: the four
scripts/sketch-prototype/*.png posters (~9.5MB, generated by the dev-only
poster.ts, not touched by sketch:check or any test) and the orphaned
.github/pr-assets/differences-page.png (~1MB). Generators kept; regenerable.
docs/pr-assets/ is retained (locked by four tests) and the DejaVu fonts are
left in place (coupled to the website build — see below).

Full unit suite green (4520 pass); no new type errors.
